refactor: site-metadata로 파일 이름 변경 (#447)
