+++ src/app/guestbook/login.tsx
@@ -1,5 +1,5 @@
 import phrases from 'data/phrases';
-import siteMetadata from 'data/siteMetadata.mjs';
+import siteMetadata from 'data/site-metadata.mjs';
 
 import { signIn } from 'next-auth/react';
 import { FcGoogle } from 'react-icons/fc';
