+++ src/components/comments/index.tsx
@@ -1,6 +1,6 @@
 import dynamic from 'next/dynamic';
 
-import siteMetadata from 'data/siteMetadata.mjs';
+import siteMetadata from 'data/site-metadata.mjs';
 
 const GiscusComponent = dynamic(() => import('./giscus'), {
   ssr: false,
