+++ src/lib/format-date.ts
@@ -1,5 +1,5 @@
 import phrases from 'data/phrases';
-import siteMetadata from 'data/siteMetadata.mjs';
+import siteMetadata from 'data/site-metadata.mjs';
 
 const OPTIONS: Intl.DateTimeFormatOptions = {
   year: 'numeric',
