+++ src/lib/scroll-pos-store.ts
@@ -1,4 +1,4 @@
-import siteMetadata from 'data/siteMetadata.mjs';
+import siteMetadata from 'data/site-metadata.mjs';
 
 interface StoreElement {
   scrollPos: number;
